\name{hfplot}
\alias{hfplot}
\title{
Plot HF data
}
\description{
Plot HF data
}
\usage{
hfplot(region = c("Alberta", "Canadian Shield", "Boreal", "Foothills", "Rocky Mountain", "Parkland", "Grassland"), type = c("Total", "Human-created Water Bodies", "Agriculture", "Urban, Rural & Industrial", "Mines, Wells & Other Energy Features", "Vegetated Linear Industrial Features", "Transportation", "Forest Harvest"))
}
\arguments{
  \item{region}{
region.
}
  \item{type}{
HF type.
}
}
\value{
Plot as a side effect.
}
\author{
Peter Solymos <solymos@ualberta.ca>
}
\examples{
hfplot()
}
\keyword{ hplot }
